Conflicts:
	vistrails/core/api.py
	vistrails/core/collection/vistrail.py
	vistrails/core/modules/module_registry.py
	vistrails/core/vistrail/controller.py
	vistrails/db/domain/__init__.py
	vistrails/db/persistence/__init__.py
	vistrails/db/services/action_chain.py
	vistrails/db/versions/__init__.py
	vistrails/gui/common_widgets.py
	vistrails/gui/module_info.py
	vistrails/gui/pipeline_view.py
	vistrails/gui/version_prop.py
	vistrails/gui/version_view.py
	vistrails/gui/vistrail_controller.py

+++ vistrails/core/thumbnails.py
@@ -1,51 +1,55 @@
 ###############################################################################
 ##
-## Copyright (C) 2011-2013, NYU-Poly.
-## Copyright (C) 2006-2011, University of Utah. 
+## Copyright (C) 2014-2015, New York University.
+## Copyright (C) 2011-2014, NYU-Poly.
+## Copyright (C) 2006-2011, University of Utah.
 ## All rights reserved.
 ## Contact: contact@vistrails.org
 ##
 ## This file is part of VisTrails.
 ##
-## "Redistribution and use in source and binary forms, with or without 
+## "Redistribution and use in source and binary forms, with or without
 ## modification, are permitted provided that the following conditions are met:
 ##
-##  - Redistributions of source code must retain the above copyright notice, 
+##  - Redistributions of source code must retain the above copyright notice,
 ##    this list of conditions and the following disclaimer.
-##  - Redistributions in binary form must reproduce the above copyright 
-##    notice, this list of conditions and the following disclaimer in the 
+##  - Redistributions in binary form must reproduce the above copyright
+##    notice, this list of conditions and the following disclaimer in the
 ##    documentation and/or other materials provided with the distribution.
-##  - Neither the name of the University of Utah nor the names of its 
-##    contributors may be used to endorse or promote products derived from 
+##  - Neither the name of the New York University nor the names of its
+##    contributors may be used to endorse or promote products derived from
 ##    this software without specific prior written permission.
 ##
-## THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" 
-## AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, 
-## THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR 
-## PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR 
-## CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, 
-## EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, 
-## PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; 
-## OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, 
-## WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR 
-## OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF 
+## THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+## AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO,
+## THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR
+## PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR
+## CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL,
+## EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
+## PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS;
+## OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY,
+## WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR
+## OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF
 ## ADVISED OF THE POSSIBILITY OF SUCH DAMAGE."
 ##
 ###############################################################################
+from __future__ import division
+
+import itertools
 
 """ Utilities for dealing with the thumbnails """
 import os
 import os.path
 import shutil
+import tempfile
 import time
 import uuid
 import mimetypes
 # mimetypes are broken by default on windows so use the builtins
 # Remove line below when it is fixed here: http://bugs.python.org/issue15207
 mimetypes.init(files=[])
-from vistrails.core import debug
-from vistrails.core.configuration import get_vistrails_configuration, \
-      get_vistrails_persistent_configuration
+from vistrails.core import debug, system
+from vistrails.core.configuration import get_vistrails_configuration
 from vistrails.core.utils import VistrailsInternalError
 
 ############################################################################
@@ -67,7 +71,13 @@ class ThumbnailCache(object):
             ThumbnailCache._instance = obj
         return ThumbnailCache._instance
 
+    @staticmethod
+    def clearInstance():
+        if ThumbnailCache._instance is not None:
+            ThumbnailCache._instance.destroy()
+
     def __init__(self):
+        self._temp_directory = None
         self.elements = {}
         self.vtelements = {}
         self.conf = None
@@ -75,17 +85,24 @@ class ThumbnailCache(object):
         if conf.has('thumbs'):
             self.conf = conf.thumbs
         self.init_cache()
+
+    def destroy(self):
+        if self._temp_directory is not None:
+            print "removing thumbnail directory"
+            shutil.rmtree(self._temp_directory)
         
     def get_directory(self):
-        if self.conf.check('cacheDirectory'):
-            thumbnail_dir = self.conf.cacheDirectory
+        thumbnail_dir = system.get_vistrails_directory('thumbs.cacheDir')
+        if thumbnail_dir is not None:
             if not os.path.exists(thumbnail_dir):
                 raise VistrailsInternalError("Cannot find %s" % thumbnail_dir)
             return thumbnail_dir
         
-        raise VistrailsInternalError("'thumbs.cacheDirectory' not"
-                                     " specified in configuration")
-        return None
+        # raise VistrailsInternalError("'thumbs.cacheDir' not"
+        #                              " specified in configuration")
+        if self._temp_directory is None:
+            self._temp_directory = tempfile.mkdtemp(prefix='vt_thumbs_')
+        return self._temp_directory
     
     def init_cache(self):
         for root,dirs, files in os.walk(self.get_directory()):
@@ -131,8 +148,9 @@ class ThumbnailCache(object):
                     entry.abs_name = dstname
                         
                 except shutil.Error, e:
-                    debug.warning("Could not move thumbnail from %s to %s: %s" \
-                                  % (sourcedir, destdir, str(e)))
+                    debug.warning("Could not move thumbnail from %s to %s" % (
+                                  sourcedir, destdir),
+                                  e)
                     
     def remove_lru(self,n=1):
         elements = self.elements.values()
@@ -141,13 +159,13 @@ class ThumbnailCache(object):
         debug.debug("Will remove %s elements from cache..."%num)
         debug.debug("Cache has %s elements and %s bytes"%(len(elements),
                                                              self.size()))
-        for i in range(num):
+        for elem in itertools.islice(elements, num):
             try:
-                del self.elements[elements[i].name]    
-                os.unlink(elements[i].abs_name)
+                del self.elements[elem.name]
+                os.unlink(elem.abs_name)
             except os.error, e:
-                debug.warning("Could not remove file %s: %s" % \
-                                 (elements[i].abs_name, str(e)))
+                debug.warning("Could not remove file %s" % elem.abs_name, e)
+
     def remove(self,key):
         if key in self.elements.keys():
             entry = self.elements[key]
@@ -176,7 +194,7 @@ class ThumbnailCache(object):
         if len(thumbnail_fnames) > 0:
             image = self._merge_thumbnails(thumbnail_fnames)
         fname = None
-        if image != None and image.width() > 0 and image.height() > 0:
+        if image is not None and image.width() > 0 and image.height() > 0:
             fname = "%s.png" % str(uuid.uuid1())
             abs_fname = self._save_thumbnail(image, fname) 
             statinfo = os.stat(abs_fname)
@@ -212,13 +230,15 @@ class ThumbnailCache(object):
         Deletes all files inside dirname
     
         """
+        if dirname is None:
+            return
         try:
             for root, dirs, files in os.walk(dirname):
                 for fname in files:
                     os.unlink(os.path.join(root,fname))
                     
         except OSError, e:
-            debug.warning("Error when removing thumbnails: %s"%str(e))
+            debug.warning("Error when removing thumbnails", e)
     
     @staticmethod
     def _get_thumbnail_fnames(folder):
